Change buildpack detection name from "Sysstat" to "Datadog"
Since whilst debugging a customer support ticket, it was confusing to see
a buildpack called "Sysstat" in the build log output, which did not appear
to match any of the buildpacks in the buildpacks URLs list for that app.

The buildpack is for Datadog, so calling it that seems more intuitive,
plus it then matches the GitHub repo URL that appears in the
buildpacks URLs list.

Example build log before:

```
-----> Deleting 1 files matching .slugignore patterns.
-----> Sysstat app detected
-----> Updating apt caches for dependencies
...
-----> Ruby app detected
...
```

Example build log after:

```
-----> Deleting 1 files matching .slugignore patterns.
-----> Datadog app detected
-----> Updating apt caches for dependencies
...
-----> Ruby app detected
...
```