Allow override of datadog config path
The default config folder path for configuration files is /datadog,
but users may wish to use a different location.
One use case is that a Rails application has a standard folder layout,
and system/infra folders tend to be dot prefixed, eg .github

This change adds a check that if DD_HEROKU_CONF_FOLDER is set, it will
use that path, otherwise it falls back to normal default of /datadog.